updates resources config to v0.2.0
This adds multi-party authorization, signoff, and composable resource
mechanisms to the config.

multi-party authorization creates a token for a resource that requires
the signoff of another authorization service before the token is valid

signoff is the configuration for the 1..n subsequent authorization
services needing to signoff on a token for a particular resource

composable resources allow a single resource to support multiple
mechanisms. for example, it is possible to create service chain tokens
with multi-party authorization.