clients.toml: adds domain-restricted identities
This adds "domain-restricted" identities and more configuration to aid
that use case.

First, what are "domain-restricted" identities?

Identities configured without a domain restriction are "realm" or
"system" identities. These identities act at the system-wide level for a
Hessra auth service and are delineated by the signature of that auth
service.

Domain or application identities, are identities that are only
applicable to one specific domain/application in the system. Another way
to put it is that a realm/system identity can be configured to mint its
own domain/application identities that can be distributed for use with
that specific domain/application.

To give a concrete example, say you have three system identities
defined: a top-level CI/CD runner, an assistant agent, and a webapp.
These each need to interact with eachother and other parts of your
system so you write policies for them to do so. But, say you want to
create new identities specifically for your webapp (like users). These
identities will ONLY be interacting with the webapp. You can configure
the webapp identity to mint domain restricted identities so it can mint
user identities to be used only with it.

On top of that, this change adds some very basic role-based access
controls. It lets you define roles and what resources they have access
to (roles are just groupings of resource+operations that are allowed to
anything assigned that role). It also lets you define a default role so
that newly minted identities can be assigned permissions by default.

To assign a domain-restricted identity to a role, you just add it as an
identity to the clients.toml